Worked on parser and fixed lexer to correctly recognize
