+++ .vscode/settings.json
@@ -1,5 +1,6 @@
 {
     "files.associations": {
-        "lexer.h": "c"
+        "lexer.h": "c",
+        "string.h": "c"
     }
 }
