+++ compiler/Makefile
@@ -1,5 +1,5 @@
-compiler: compiler.c lexer.c
-	gcc -o compiler compiler.c lexer.c
+compiler: compiler.c lexer.c parser.c
+	gcc -o compiler compiler.c lexer.c parser.c
 
 test: test.c lexer.c
 	gcc -o test test.c lexer.c
