+++ compiler/compiler.c
@@ -6,20 +6,8 @@
 
 #include "types.h"
 #include "lexer.h"
-
-
-void freeTokens(token *tokens, int n);
-token *nextToken();
-
-program *parseProgram();
-function *parseFunc();
-statement *parseStatement();
-expression *parseExpression();
-
-void freeProgram(program *p);
-void freeFunc(function *f);
-void freeStatement(statement *s);
-void freeExpression(expression *e);
+#include "parser.h"
+#include "tokens.h"
 
 
 
@@ -60,7 +48,7 @@ int main(int argc, char *argv[])
 
    
 
-    printf("\n");
+    //printf("\n");
     // Print tokens
     for (int i = 0; i < num_tokens; i++)
     {
@@ -96,131 +84,6 @@ void freeTokens(token *tokens, int num_tokens)
     }
 
     free (tokens);
-}
-
-program *parseProgram()
-{
-    program *p;
-    p->func = parseFunc();
-    return p;
-}
-
-function *parseFunc()
-{
-    function *f = malloc(sizeof(function));
-    token *tok = nextToken();
-    // For now functions can only be of type INT
-    if (tok == NULL || tok->type != KEYWORD_INT)
-    {
-        printf("Missing function return type\n");
-        exit(1);
-    }
-
-    // Look for and set function ID
-    tok = nextToken();
-    if (tok == NULL || tok->type != IDENTIFIER)
-    {
-        printf("Missing function name\n");
-        exit(1);
-    }
-    f->id = tok->lexeme;
-
-    // Parentheses
-    tok = nextToken();
-    if (tok == NULL || tok->type != OPEN_PAR)
-    {
-        printf("Missing parentheses\n");
-        exit(1);
-    }
-    tok = nextToken();
-    if (tok == NULL || tok->type != CLOSED_PAR)
-    {
-        printf("Missing parentheses\n");
-        exit(1);
-    }
-
-    // Function body
-    tok = nextToken();
-    if (tok == NULL || tok->type != OPEN_BRACE)
-    {
-        printf("Missing bracket\n");
-        exit(1);
-    }
-
-    // Parse statement
-    f->statement = parseStatement();
-    tok = nextToken();
-
-
-    if (tok == NULL || tok->type != CLOSED_BRACE)
-    {
-        printf("Missing bracket\n");
-        exit(1);
-    }
-    return f;
-}
-
-statement *parseStatement()
-{
-    statement *s = malloc(sizeof(statement));
-
-    // Check for return statement
-    token *tok = nextToken();
-    if (tok == NULL || tok->type != RETURN)
-    {
-        printf("Missing return statement\n");
-        exit(1);
-    }
-
-    // Get expression
-    s->exp = parseExpression();
-
-    //Check for a semicolon
-    if (tok == NULL || tok->type != SEMICOLON)
-    {
-        printf("Missing Semicolon\n");
-        exit(1);
-    }
-    return s;
-}
-
-expression *parseExpression()
-{
-    expression *e = malloc(sizeof(expression));
-
-    // Check for int value
-    token *tok = nextToken();
-    if (tok == NULL || tok->type != INT)
-    {
-        printf("Missing expression");
-        exit(1);
-    }
-    e->value = tok->value;
-    return e;
-}
-
-void freeProgram(program *p)
-{
-    freeFunc(p->func);
-    free(p);
-}
-
-void freeFunc(function *f)
-{
-    free(f->id);
-    freeStatement(f->statement);
-    free(f);
-}
-
-void freeStatement(statement *s)
-{
-    freeExpression(s->exp);
-    free(s);
-}
-
-void freeExpression(expression *e)
-{
-    free(e);
 }
 
 // Returns a pointer to the next token from the token list
