+++ compiler/tokens.h
@@ -1,0 +1,9 @@
+#ifndef TOKENS_H
+#define TOKENS_H
+
+#include "types.h"
+
+void freeTokens(token *tokens, int n);
+token *nextToken();
+
+#endif
